add a test for updateMembership
